pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-roaster"
version = "0.1.0"
description = "A CLI tool that takes a code file and roasts it using AI"
readme = "README.md"
authors = [
    {name = "Dan Barr"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "langchain-anthropic>=0.0.1",
    "langchain-community>=0.0.1",
    "langchain-ollama>=0.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.3",
    "rich>=13.3.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
code-roaster = "code_roaster.cli:main"

[tool.setuptools]
packages = ["code_roaster"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88